feat: add unset-properties command + flags refactoring
